Technical/Move exceptions to Error namespace
* Moved exceptions to DnsMock::Error namespace
* Moved error helper to DnsMock::Error namespace
