+++ lib/dns_mock/error/argument_type.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module DnsMock
+  module Error
+    ArgumentType = ::Class.new(::StandardError) do
+      def initialize(argument_class)
+        super("Argument class is a #{argument_class}. Should be a Hash")
+      end
+    end
+  end
+end
