+++ lib/dns_mock/error/record_not_found.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module DnsMock
+  module Error
+    RecordNotFound = ::Class.new(::StandardError) do
+      def initialize(record_type, hostname)
+        super("#{record_type} not found for #{hostname} in predefined records dictionary")
+      end
+    end
+  end
+end
