+++ lib/dns_mock/server/records_dictionary_builder.rb
@@ -3,7 +3,7 @@
 module DnsMock
   class Server
     class RecordsDictionaryBuilder
-      include DnsMock::Helper::Error
+      include DnsMock::Error::Helper
 
       TYPE_MAPPER = DnsMock::AVAILABLE_DNS_RECORD_TYPES.zip(
         [
@@ -28,7 +28,7 @@ module DnsMock
       end
 
       def build(records_to_build)
-        raise DnsMock::ArgumentTypeError, records_to_build.class unless records_to_build.is_a?(::Hash)
+        raise DnsMock::Error::ArgumentType, records_to_build.class unless records_to_build.is_a?(::Hash)
 
         records_to_build.each do |hostname, dns_records|
           records[hostname] = dns_records.each_with_object({}) do |(record_type, records_data), records_instances_by_type|
