+++ spec/dns_mock/error/record_context_type_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+RSpec.describe DnsMock::Error::RecordContextType do
+  subject(:error_instance) { described_class.new('record_context_type', 'record_type', 'expected_type') }
+
+  let(:error_context) { 'record_context_type is invalid record context type for record_type record. Should be a expected_type' }
+
+  it_behaves_like 'customized standard error'
+end
