+++ spec/support/shared_examples/target_class_exception_wrapper.rb
@@ -3,7 +3,7 @@
 module DnsMock
   RSpec.shared_examples 'target class exception wrapper' do
     it 'wraps target class exception and raises as customized error' do
-      expect { create_factory }.to raise_error(DnsMock::RecordContextError, error_context)
+      expect { create_factory }.to raise_error(DnsMock::Error::RecordContext, error_context)
     end
   end
 end
